refactor: group shared form components
