+++ src/shared/components/forms/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Button as Template } from "@src/shared/components/Button";
+import { Button as Template } from "@src/shared/components/forms/Button";
 
 const meta = {
   title: "Shared/Button",
