+++ src/shared/components/forms/Form.stories.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Form as Template } from "@src/shared/components/Form";
-import { FormItem } from "@src/shared/components/FormItem";
+import { Form as Template } from "@src/shared/components/forms/Form";
+import { FormItem } from "@src/shared/components/forms/FormItem";
 
 const meta = {
   title: "Shared/Form",
