+++ src/shared/components/forms/FormItem.stories.tsx
@@ -2,12 +2,12 @@ import React from "react";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Button } from "@src/shared/components/Button";
-import { FormItem as Template } from "@src/shared/components/FormItem";
-import { Input } from "@src/shared/components/Input";
-import { Select } from "@src/shared/components/Select";
-import { Slider } from "@src/shared/components/Slider";
-import { Switch } from "@src/shared/components/Switch";
+import { Button } from "@src/shared/components/forms/Button";
+import { FormItem as Template } from "@src/shared/components/forms/FormItem";
+import { Input } from "@src/shared/components/forms/Input";
+import { Select } from "@src/shared/components/forms/Select";
+import { Slider } from "@src/shared/components/forms/Slider";
+import { Switch } from "@src/shared/components/forms/Switch";
 
 const meta = {
   title: "Shared/FormItem",
