+++ src/shared/components/forms/Select.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Select as Template } from "@src/shared/components/Select";
+import { Select as Template } from "@src/shared/components/forms/Select";
 import * as fixture from "@src/shared/storybook/fixture";
 
 const meta = {
