+++ src/shared/components/forms/Slider.tsx
