+++ src/shared/components/index.ts
@@ -1,13 +1,13 @@
-export * from "@src/shared/components/Button";
+export * from "@src/shared/components/forms/Button";
 export * from "@src/shared/components/Card";
 export * from "@src/shared/components/Code";
 export * from "@src/shared/components/Description";
 export * from "@src/shared/components/Feedback";
-export * from "@src/shared/components/Form";
-export * from "@src/shared/components/FormItem";
+export * from "@src/shared/components/forms/Form";
+export * from "@src/shared/components/forms/FormItem";
 export * from "@src/shared/components/layout/FullScreen";
 export * from "@src/shared/components/layout/Header";
-export * from "@src/shared/components/Input";
+export * from "@src/shared/components/forms/Input";
 export * from "@src/shared/components/layout/Layout";
 export * from "@src/shared/components/layout/List";
 export * from "@src/shared/components/Logo";
@@ -17,12 +17,12 @@ export * from "@src/shared/components/layout/Outer";
 export * from "@src/shared/components/Overlay";
 export * from "@src/shared/components/Score";
 export * from "@src/shared/components/Section";
-export * from "@src/shared/components/Select";
-export * from "@src/shared/components/Slider";
+export * from "@src/shared/components/forms/Select";
+export * from "@src/shared/components/forms/Slider";
 export * from "@src/shared/components/Style";
-export * from "@src/shared/components/Switch";
-export * from "@src/shared/components/Tag";
+export * from "@src/shared/components/forms/Switch";
+export * from "@src/shared/components/forms/Tag";
 export * from "@src/shared/components/TagList";
-export * from "@src/shared/components/Textarea";
+export * from "@src/shared/components/forms/Textarea";
 export * from "@src/shared/components/Title";
-export * from "@src/shared/components/Upload";
+export * from "@src/shared/components/forms/Upload";
